Planted the seeds of UI rendering
Wow, what a pain that was.  Lots of stuff to get right.
However, I now have what appears to be a fairly nice way
of dealing with secondary command buffers.

While I was at it, I did some cleanup.  For example, I
got that whole number of swapchain images vs. number of
in-flight frames thing sorted out, and I think I understand
the synchronization requirements of acquiring an image,
rendering to it, and presenting it much better than I did before.